refactor(style): extract Paint<TextEdit> into default_style/text_edit.rs
Convert default_style.rs to a directory module (default_style/mod.rs) and
lift the Paint<TextEdit> impl into a sibling default_style/text_edit.rs.
Add quartzite-geometry as a production dep to quartzite-style (needed by the
paint helpers being added in subtask 9). Mark helper functions pub(super) so
text_edit.rs can import them via super::. The #[path] for the test module
is updated to "../default_style_tests.rs" to resolve from the new directory.
mod.rs shrinks by ~59 lines; total file size headroom is now available for
the ~150-line growth from the caret/selection paint helpers in subtask 9.
